refactor(sessions): move lifecycle snapshot equality
